+++ qa/sc_evm_mc_addr_ownership.py
@@ -240,7 +240,7 @@ class SCEvmMcAddressOwnership(AccountChainSetup):
             generate_next_block(sc_node, "first node", force_switch_to_next_epoch=True)
             self.sc_sync_all()
 
-        # add the same sc/mc ownership , as abovesending a transaction with data invoking native smart contract
+        # add the same sc/mc ownership , as above sending a transaction with data invoking native smart contract
         ret = sendKeysOwnership(sc_node, nonce=1,
                                 sc_address=sc_address,
                                 mc_addr=taddr1,